+++ tests/Database/Adapter/PgsqlAdapterTest.php
@@ -53,7 +53,7 @@ class PgsqlAdapterTest extends TestCase
         $queryBuilder = $this->adapter->getQueryBuilder();
         $queries = $queryBuilder->createTable($migrationTable);
         foreach ($queries as $query) {
-            $this->adapter->execute($query);
+            $this->adapter->query($query);
         }
 
         $updatedStructure = $this->adapter->getStructure();
@@ -313,7 +313,7 @@ class PgsqlAdapterTest extends TestCase
         $migrationTable1->create();
         $queries1 = $queryBuilder->createTable($migrationTable1);
         foreach ($queries1 as $query) {
-            $this->adapter->execute($query);
+            $this->adapter->query($query);
         }
 
         $migrationTable2 = new MigrationTable('table_2');
@@ -355,7 +355,7 @@ class PgsqlAdapterTest extends TestCase
         $migrationTable2->create();
         $queries2 = $queryBuilder->createTable($migrationTable2);
         foreach ($queries2 as $query) {
-            $this->adapter->execute($query);
+            $this->adapter->query($query);
         }
     }