+++ tests/Helpers/Adapter/MysqlCleanupAdapter.php
@@ -9,7 +9,7 @@ class MysqlCleanupAdapter extends MysqlAdapter implements CleanupInterface
     public function cleanupDatabase()
     {
         $database = getenv('PHOENIX_MYSQL_DATABASE');
-        $this->execute(sprintf('DROP DATABASE IF EXISTS `%s`', $database));
-        $this->execute(sprintf('CREATE DATABASE `%s`', $database));
+        $this->query(sprintf('DROP DATABASE IF EXISTS `%s`', $database));
+        $this->query(sprintf('CREATE DATABASE `%s`', $database));
     }
 }